refactor(split): Task 14.0 前置 —— api.ts/renderer-recovery.ts 拆出 600 行约束内子模块
- lib/logger/api.ts 685→480 行，诊断 zip（AC-8）拆至 diagnostics.ts
- renderer-recovery.ts 677→门面 27 行 + policy/load/machine 三模块（公开面不变）
- electron-builder.yml files 清单补拆分子目录条目
- bundled-files.test.ts 新增入口 require 闭包遍历测试（main/host-bootstrap/preload 出发，防门面在清单、实现漏包的启动闪退）
- 全量 492/492 + typecheck 零错 + pack 产物核验通过
